+++ packages/data/addon-test-support/mirage/scenarios/elide-one.ts
@@ -1,16 +1,19 @@
-// TODO: Replace any with type supplied by new version of mirage
-import { grains } from './elide-one';
+/**
+ * Copyright 2020, Yahoo Holdings Inc.
+ * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
+ */
 import { capitalize } from '@ember/string';
 
+export const grains = ['hour', 'day', 'week', 'month', 'quarter', 'year'];
 const timeDimIds = ['eventTime', 'orderTime'];
 
+// TODO: Replace any with type supplied by new version of mirage
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export default function(server: any) {
   const [table0, table1] = server.createList('table', 2);
-  server.createList('metric', 1, { table: table0 });
+  server.createList('metric', 3, { table: table0 });
   server.createList('metric', 2, { table: table1 });
-  server.createList('dimension', 2, { table: table0 });
-  server.createList('dimension', 4, { table: table1 });
+  server.createList('dimension', 3, { table: table0 });
   const timeDimTables = [table1];
   grains.forEach(grain => {
     timeDimIds.forEach(timeDimId => {
