+++ packages/data/tests/dummy/mirage/config.js
@@ -1,6 +1,6 @@
 import config from 'dummy/config/environment';
 import BardMeta from 'navi-data/mirage/routes/bard-meta';
-import GraphQL from 'navi-data/mirage/routes/graphql';
+import GraphQL from 'navi-data/test-support/mirage/routes/graphql';
 
 export default function() {
   this.passthrough('/write-coverage');
