+++ packages/data/tests/dummy/mirage/factories/dimension.js
@@ -1,1 +1,1 @@
-export { default } from 'navi-data/mirage/factories/dimension';
+export { default } from 'navi-data/test-support/mirage/factories/dimension';
