+++ packages/data/tests/dummy/mirage/factories/table.js
@@ -1,1 +1,1 @@
-export { default } from 'navi-data/mirage/factories/table';
+export { default } from 'navi-data/test-support/mirage/factories/table';
