+++ packages/data/tests/unit/models/metadata/metric-test.ts
@@ -3,7 +3,7 @@ import { setupTest } from 'ember-qunit';
 import Pretender, { Server } from 'pretender';
 import { TestContext } from 'ember-test-helpers';
 //@ts-ignore
-import metadataRoutes from '../../../helpers/metadata-routes';
+import metadataRoutes from 'navi-data/test-support/helpers/metadata-routes';
 import MetricMetadataModel, { MetricMetadataPayload } from 'navi-data/models/metadata/metric';
 import ColumnFunctionMetadataModel from 'navi-data/models/metadata/column-function';
