+++ packages/data/tests/unit/services/metric-parameter-test.js
@@ -3,7 +3,7 @@ import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 import config from 'ember-get-config';
 import Pretender from 'pretender';
-import metadataRoutes from '../../helpers/metadata-routes';
+import metadataRoutes from 'navi-data/test-support/helpers/metadata-routes';
 import { groupBy } from 'lodash-es';
 
 const HOST = config.navi.dataSources[0].uri;
