+++ packages/data/tsconfig.json
@@ -36,7 +36,6 @@
     "addon/**/*",
     "tests/**/*",
     "types/**/*",
-    "test-support/**/*",
     "addon-test-support/**/*",
     "../*/config/*"
   ]
